# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentResponse": {
        "properties": {
          "assignee_name": {
            "title": "Assignee Name",
            "type": "string"
          },
          "is_override": {
            "title": "Is Override",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/AssignmentRole"
          },
          "rule_violations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RuleViolationResponse"
            },
            "title": "Rule Violations",
            "type": "array"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          }
        },
        "required": [
          "service_date",
          "role",
          "assignee_name",
          "is_override"
        ],
        "title": "AssignmentResponse",
        "type": "object"
      },
...
      "BatchOverrideRequest": {
        "additionalProperties": false,
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/BatchOverrideItem"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Assignments",
            "type": "array"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "schedule_id": {
            "format": "uuid",
            "title": "Schedule Id",
            "type": "string"
          }
        },
        "required": [
          "schedule_id",
          "expected_version",
          "assignments",
          "reason"
        ],
        "title": "BatchOverrideRequest",
        "type": "object"
      },
...
      "CalendarAssignmentResponse": {
        "properties": {
          "assignee_name": {
            "title": "Assignee Name",
            "type": "string"
          },
          "change_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Kind"
          },
          "is_override": {
            "title": "Is Override",
            "type": "boolean"
          },
          "member_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "role": {
            "$ref": "#/components/schemas/AssignmentRole"
          },
          "schedule_id": {
            "format": "uuid",
            "title": "Schedule Id",
            "type": "string"
          },
          "schedule_version": {
            "title": "Schedule Version",
            "type": "integer"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          }
        },
        "required": [
          "schedule_id",
          "schedule_version",
          "service_date",
          "role",
          "assignee_name",
          "member_id",
          "is_override",
          "change_kind"
        ],
        "title": "CalendarAssignmentResponse",
        "type": "object"
      },
...
      "CalendarEventCreate": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "$ref": "#/components/schemas/CalendarEventColor"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "starts_on",
          "ends_on",
          "title",
          "color"
        ],
        "title": "CalendarEventCreate",
        "type": "object"
      },
...
      "CalendarEventResponse": {
        "properties": {
          "color": {
            "$ref": "#/components/schemas/CalendarEventColor"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "color",
          "starts_on",
          "ends_on",
          "created_at"
        ],
        "title": "CalendarEventResponse",
        "type": "object"
      },
      "CalendarEventUpdate": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarEventColor"
              },
              {
                "type": "null"
              }
            ]
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CalendarEventUpdate",
        "type": "object"
      },
...
      "CalendarResponse": {
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/CalendarAssignmentResponse"
            },
            "title": "Assignments",
            "type": "array"
          },
          "availability": {
            "items": {
              "$ref": "#/components/schemas/CalendarAvailabilityResponse"
            },
            "title": "Availability",
            "type": "array"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/CalendarDayResponse"
            },
            "title": "Days",
            "type": "array"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/CalendarMemberResponse"
            },
            "title": "Members",
            "type": "array"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          }
        },
        "required": [
          "starts_on",
          "ends_on",
          "days",
          "members",
          "assignments",
          "availability"
        ],
        "title": "CalendarResponse",
        "type": "object"
      },
...
      "DirectOverrideCheckRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "replacement_member_id": {
            "format": "uuid",
            "title": "Replacement Member Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AssignmentRole"
          },
          "schedule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          }
        },
        "required": [
          "service_date",
          "role",
          "replacement_member_id"
        ],
        "title": "DirectOverrideCheckRequest",
        "type": "object"
      },
      "DirectOverrideRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "replacement_member_id": {
            "format": "uuid",
            "title": "Replacement Member Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AssignmentRole"
          },
          "schedule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "service_date",
          "role",
          "replacement_member_id"
        ],
        "title": "DirectOverrideRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleViolationResponse": {
        "description": "One hard rule an edit would break (decision D3).",
        "properties": {
          "days": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Days",
            "type": "array"
          },
          "member_name": {
            "title": "Member Name",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          }
        },
        "required": [
          "rule",
          "message",
          "member_name",
          "days"
        ],
        "title": "RuleViolationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Erste On-call",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/calendar": {
      "get": {
        "operationId": "calendar_matrix_api_v1_calendar_get",
        "parameters": [
          {
            "in": "query",
            "name": "starts_on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Starts On",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ends_on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Ends On",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calendar Matrix",
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/events": {
      "get": {
        "operationId": "list_calendar_events_api_v1_calendar_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "starts_on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Starts On",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ends_on",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Ends On",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarEventResponse"
                  },
                  "title": "Response List Calendar Events Api V1 Calendar Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Calendar Events",
        "tags": [
          "calendar"
        ]
      },
      "post": {
        "operationId": "create_calendar_event_api_v1_calendar_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Calendar Event",
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/events/{event_id}": {
      "delete": {
        "operationId": "delete_calendar_event_api_v1_calendar_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Calendar Event",
        "tags": [
          "calendar"
        ]
      },
      "patch": {
        "operationId": "update_calendar_event_api_v1_calendar_events__event_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Calendar Event",
        "tags": [
          "calendar"
        ]
      }
    },
...
    "/api/v1/calendar/override": {
      "post": {
        "operationId": "direct_override_api_v1_calendar_override_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Direct Override",
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/override/batch": {
      "post": {
        "operationId": "batch_override_api_v1_calendar_override_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssignmentResponse"
                  },
                  "title": "Response Batch Override Api V1 Calendar Override Batch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Override",
        "tags": [
          "calendar"
        ]
      }
    },
    "/api/v1/calendar/override/check": {
      "post": {
        "description": "The hard rules the override would break, computed before the fact so\nthe confirmation dialog can show them (decision D3: warned, not blocked).",
        "operationId": "direct_override_check_api_v1_calendar_override_check_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "oncall_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oncall Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectOverrideCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleViolationResponse"
                  },
                  "title": "Response Direct Override Check Api V1 Calendar Override Check Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Direct Override Check",
        "tags": [
          "calendar"
        ]
      }
    },
...
  }
}